backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.2
openai==1.51.0
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv==1.0.1
